pyproject.toml
[project]
name = "open-books"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "docling>=2.68.0",
  "fastapi[all]>=0.128.0",
  "ipykernel>=7.1.0",
  "loguru>=0.7.3",
  "marker-pdf>=1.10.1",
  "pydantic-settings>=2.12.0",
  "pymupdf>=1.26.7",
  "pytest>=9.0.2",
  "pyyaml>=6.0.3",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
markers = ["slow: marks tests as slow (deselect with '-m \"not slow\"')"]

[dependency-groups]
dev = ["pytest-mock>=3.15.1"]
